Update build to use RPCEmu 0.9.4; move the mapping to /riscos, etc.
The RPCEmu build that we use is now 0.9.4.
The mapping of the volumes is now a lot easier to understand, with the
RISC OS disc living at /riscos, and RPCEmu living at /rpcemu.

This means that remapping the entire directory structure is significantly
easier.

The fluxbox configuration is also simplified to remove most of the menus
and the task bar. This should mean that we are not distracted from the
application.