Add module which tracks mode changes to trigger a resize.
The module watches for a change in the size of the screen through a
mode change. When one happens it updates a file in the root of the
HostFS disc. This file can then be seen by a small background process
and cause the VNC view to be resized to reflect the new window size.